chore: Factor out archive reading
